# scripts/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.1.0",
    "title": "Houses of the Oireachtas Open Data APIs",
...
    "license": {
      "name": "Oireachtas (Open Data) PSI Licence ",
      "url": "https://www.oireachtas.ie/en/open-data/license/"
    },
    "contact": {
      "name": "Oireachtas Open Data",
      "email": "open.data@oireachtas.ie",
      "url": "https://www.oireachtas.ie/en/open-data"
    }
  },
  "host": "api.oireachtas.ie",
  "basePath": "/v1",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/constituencies": {
      "x-swagger-router-controller": "constituencies",
      "get": {
        "summary": "Constituencies Endpoint",
        "description": "Returns list of constituencies filtered by the base and additional parameters.\nThe list supports paging.\n\n#### Mapping\n* chamber_id - house.uri\n* chamber - house.houseCode\n* house_no - house.houseNo\n* skip - this will ignore the first x number of records set in the parameter\n* limit - this will only return a specific amount of records",
        "tags": [
          "constituencies"
        ],
        "operationId": "constituencies",
        "parameters": [
          {
            "$ref": "#/parameters/chamberID"
          },
          {
            "$ref": "#/parameters/house"
          },
          {
            "$ref": "#/parameters/houseNo"
          },
          {
            "$ref": "#/parameters/skip"
          },
          {
            "$ref": "#/parameters/limit"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "head": {
                  "type": "object",
                  "properties": {
                    "counts": {
                      "type": "object",
                      "properties": {
                        "constituencyCount": {
                          "type": "integer",
                          "example": 8
                        },
                        "resultCount": {
                          "type": "integer",
                          "example": 8
                        }
                      }
                    }
                  }
                },
                "results": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "constituencyOrPanel": {
                        "type": "object",
                        "properties": {
                          "uri": {
                            "type": "string",
                            "format": "uri",
                            "example": "https://data.oireachtas.ie/ie/oireachtas/house/seanad/9/panel/Nominated-by-the-Taoiseach"
                          },
                          "representType": {
                            "type": "string",
                            "example": "panel"
                          },
                          "representCode": {
                            "type": "string",
                            "example": "Nominated-by-the-Taoiseach"
                          },
                          "showAs": {
                            "type": "string",
                            "example": "Nominated by the Taoiseach"
                          }
                        }
                      },
                      "house": {
                        "type": "object",
                        "properties": {
                          "houseNo": {
                            "type": "string",
                            "example": "9"
                          },
                          "uri": {
                            "type": "string",
                            "format": "uri",
                            "example": "https://data.oireachtas.ie/ie/oireachtas/house/seanad/9"
                          },
                          "houseCode": {
                            "type": "string",
                            "example": "seanad"
                          },
                          "showAs": {
                            "type": "string",
                            "example": "9th Seanad"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
    "/members": {
      "get": {
        "summary": "Members Endpoint",
        "description": "Returns a list of members.\nThe list supports paging.\n\n#### Mapping\n* date_start (Greater Than or Equal To) - member.memberships.membership.dateRange.start\n* date_end (Less Than or Equal To) - member.memberships.membership.dateRange.start\n* chamber_id - member.memberships.membership.house.uri\n* chamber - member.memberships.membership.house.houseCode\n* house_no - member.memberships.membership.house.houseNo\n* member_id - member.uri\n* skip - this will ignore the first x number of records set in the parameter\n* limit - this will only return a specific amount of records\n* party_code - member.memberships.membership.parties.party.partyCode\n* party_id - member.memberships.membership.parties.party.uri\n* const_code - member.memberships.membership.represents.represent.representCode\n* const_id - member.memberships.membership.represents.represent.uri\n* fuzzy_name_search - member.fullName",
        "tags": [
          "members"
        ],
        "operationId": "members",
        "parameters": [
          {
            "$ref": "#/parameters/dateStart"
          },
          {
            "$ref": "#/parameters/dateEnd"
          },
          {
            "$ref": "#/parameters/chamberID"
          },
          {
            "$ref": "#/parameters/house"
          },
          {
            "$ref": "#/parameters/houseNo"
          },
          {
            "$ref": "#/parameters/memberID"
          },
          {
            "$ref": "#/parameters/skip"
          },
          {
            "$ref": "#/parameters/limit"
          },
          {
            "$ref": "#/parameters/partyCode"
          },
          {
            "$ref": "#/parameters/partyID"
          },
          {
            "$ref": "#/parameters/constituencyCode"
          },
          {
            "$ref": "#/parameters/constituencyID"
          },
          {
            "$ref": "#/parameters/fuzzyNameSearch"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "head": {
                  "type": "object",
                  "properties": {
                    "counts": {
                      "type": "object",
                      "properties": {
                        "memberCount": {
                          "type": "integer",
                          "example": 1927
                        },
                        "resultCount": {
                          "type": "integer",
                          "example": 1927
                        }
                      }
                    }
                  }
                },
                "results": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "member": {
                        "type": "object",
                        "properties": {
                          "gender": { "type": "string" },
                          "uri": { "type": "string", "format": "uri" },
                          "pId": { "type": "string" },
                          "firstName": { "type": "string" },
                          "lastName": { "type": "string" },
                          "image": { "type": "boolean" },
                          "wikiTitle": { "type": "string" },
                          "memberships": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "membership": {
                                  "type": "object",
                                  "properties": {
                                    "house": {
                                      "type": "object",
                                      "properties": {
                                        "uri": { "type": "string", "format": "uri" },
                                        "houseCode": { "type": "string" },
                                        "houseNo": { "type": "string" },
                                        "chamberType": { "type": "string" },
                                        "showAs": { "type": "string" }
                                      }
                                    },
                                    "represents": {
                                      "type": "array",
                                      "items": {
                                        "type": "object",
                                        "properties": {
                                          "represent": {
                                            "type": "object",
                                            "properties": {
                                              "representCode": { "type": "string" },
                                              "uri": { "type": "string", "format": "uri" },
                                              "representType": { "type": "string" },
                                              "showAs": { "type": "string" }
                                            }
                                          }
                                        }
                                      }
                                    },
                                    "uri": { "type": "string", "format": "uri" },
                                    "dateRange": {
                                      "type": "object",
                                      "properties": {
                                        "start": { "type": "string", "format": "date" },
                                        "end": { "type": "string", "format": "date" }
                                      }
                                    },
                                    "committees": {
                                      "type": "array",
                                      "items": {}
                                    },
                                    "parties": {
                                      "type": "array",
                                      "items": {
                                        "type": "object",
                                        "properties": {
                                          "party": {
                                            "type": "object",
                                            "properties": {
                                              "partyCode": { "type": "string" },
                                              "uri": { "type": "string", "format": "uri" },
                                              "dateRange": {
                                                "type": "object",
                                                "properties": {
                                                  "start": { "type": "string", "format": "date" },
                                                  "end": { "type": "string", "format": "date" }
                                                }
                                              },
                                              "showAs": { "type": "string" }
                                            }
                                          }
                                        }
                                      }
                                    },
                                    "offices": {
                                      "type": "array",
                                      "items": {}
                                    }
                                  }
                                }
                              }
                            }
                          },
                          "dateOfDeath": { "type": "string", "format": "date" },
                          "fullName": { "type": "string" },
                          "memberCode": { "type": "string" },
                          "showAs": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/parties": {
      "x-swagger-router-controller": "parties",
      "get": {
        "summary": "Parties Endpoint",
        "description": "Returns list of parties filtered by the base and additional parameters.\nThe list supports paging.\n#### Mapping\n* chamber_id - house.uri\n* chamber - house.houseCode\n* house_no - house.houseNo\n* skip - this will ignore the first x number of records set in the parameter\n* limit - this will only return a specific amount of records",
        "tags": [
          "parties"
        ],
        "operationId": "parties",
        "parameters": [
          {
            "$ref": "#/parameters/chamberID"
          },
          {
            "$ref": "#/parameters/house"
          },
          {
            "$ref": "#/parameters/houseNo"
          },
          {
            "$ref": "#/parameters/skip"
          },
          {
            "$ref": "#/parameters/limit"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of parties and their associated house",
            "schema": {
              "type": "object",
              "properties": {
                "head": {
                  "type": "object",
                  "properties": {
                    "counts": {
                      "type": "object",
                      "properties": {
                        "partyCount": {
                          "type": "integer",
                          "example": 11
                        },
                        "resultCount": {
                          "type": "integer",
                          "example": 11
                        }
                      }
                    }
                  }
                },
                "results": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "party": {
                        "type": "object",
                        "properties": {
                          "partyCode": {
                            "type": "string",
                            "example": "Sinn_FÃ©in"
                          },
                          "uri": {
                            "type": "string",
                            "format": "uri",
                            "example": "https://data.oireachtas.ie/ie/oireachtas/party/dail/31/Sinn_FÃ©in"
                          },
                          "showAs": {
                            "type": "string",
                            "example": "Sinn FÃ©in"
                          }
                        }
                      },
                      "house": {
                        "type": "object",
                        "properties": {
                          "showAs": {
                            "type": "string",
                            "example": "31st DÃ¡il"
                          },
                          "uri": {
                            "type": "string",
                            "format": "uri",
                            "example": "https://data.oireachtas.ie/ie/oireachtas/house/dail/31"
                          },
                          "houseCode": {
                            "type": "string",
                            "example": "dail"
                          },
                          "houseNo": {
                            "type": "string",
                            "example": "31"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/questions": {
      "x-swagger-router-controller": "questions",
      "get": {
        "summary": "Questions Endpoint",
        "description": "Returns list of questions filtered by the base and additional parameters.\nThe list supports paging.\n#### Mapping\n* date_start (Greater Than or Equal To) - question.date\n* date_end (Less Than or Equal To) - question.date\n* skip - this will ignore the first x number of records set in the parameter\n* limit - this will only return a specific amount of records\n* qtype - question.questionType\n* member_id - question.by.uri\n* question_id - question.uri\n* question_no - question.questionNumber",
        "tags": [
          "questions"
        ],
        "operationId": "questions",
        "parameters": [
          {
            "$ref": "#/parameters/dateStart"
          },
          {
            "$ref": "#/parameters/dateEnd"
          },
          {
            "$ref": "#/parameters/skip"
          },
          {
            "$ref": "#/parameters/limit"
          },
          {
            "$ref": "#/parameters/questionType"
          },
          {
            "$ref": "#/parameters/memberID"
          },
          {
            "$ref": "#/parameters/questionID"
          },
          {
            "$ref": "#/parameters/questionNo"
          },
          {
            "$ref": "#/parameters/showAnswers"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "head": {
                  "type": "object",
                  "properties": {
                    "counts": {
                      "type": "object",
                      "properties": {
                        "questionCount": {
                          "type": "integer",
                          "example": 10000
                        },
                        "resultCount": {
                          "type": "integer",
                          "example": 10000
                        }
                      }
                    }
                  }
                },
                "results": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "question": {
                        "type": "object",
                        "properties": {
                          "answerText": { "type": "string" },
                          "questionNumber": { "type": "integer" },
                          "by": {
                            "type": "object",
                            "properties": {
                              "showAs": { "type": "string" },
                              "memberCode": { "type": "string" },
                              "uri": { "type": "string", "format": "uri" }
                            }
                          },
                          "to": {
                            "type": "object",
                            "properties": {
                              "showAs": { "type": "string" },
                              "uri": { "type": "string", "format": "uri" },
                              "roleCode": { "type": "string" },
                              "roleType": { "type": "string" }
                            }
                          },
                          "house": {
                            "type": "object",
                            "properties": {
                              "uri": { "type": "string", "format": "uri" },
                              "houseCode": { "type": "string" },
                              "houseNo": { "type": "string" },
                              "showAs": { "type": "string" },
                              "chamberType": { "type": "string" },
                              "committeeCode": { "type": "string" }
                            }
                          },
                          "uri": { "type": "string", "format": "uri" },
                          "showAs": { "type": "string" },
                          "debateSection": {
                            "type": "object",
                            "properties": {
                              "showAs": { "type": "string" },
                              "uri": { "type": "string", "format": "uri" },
                              "formats": {
                                "type": "object",
                                "properties": {
                                  "pdf": {
                                    "type": "object",
                                    "properties": {
                                      "uri": { "type": "string", "format": "uri" }
                                    }
                                  },
                                  "xml": {
                                    "type": "object",
                                    "properties": {
                                      "uri": { "type": "string", "format": "uri" }
                                    }
                                  }
                                }
                              },
                              "debateSectionId": { "type": "string" }
                            }
                          },
                          "date": { "type": "string", "format": "date" },
                          "questionType": { "type": "string" }
                        }
                      },
                      "contextDate": { "type": "string", "format": "date" }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "ErrorResponse": {
      "required": [
        "message"
      ],
      "properties": {
        "message": {
          "type": "string"
        }
      }
    }
  },
  "parameters": {
...
    "questionType": {
      "name": "qtype",
      "in": "query",
      "description": "Filter questions by oral or writtens.",
      "required": false,
      "type": "array",
      "items": {
        "type": "string",
        "enum": [
          "oral",
          "written"
        ]
      },
      "collectionFormat": "csv",
      "default": [
        "oral",
        "written"
      ],
      "maxItems": 2,
      "minItems": 0
    },
    "showAnswers": {
      "name": "show_answers",
      "in": "query",
      "description": "Add answers to the questions response",
      "required": false,
      "type": "boolean"
    },
    "memberID": {
      "name": "member_id",
      "in": "query",
      "description": "Filter by Member uri, e.g. /ie/oireachtas/member/id/Brendan-Howlin.S.1983-02-23",
      "required": false,
      "type": "string"
    },
...
    "house": {
      "name": "chamber",
      "in": "query",
      "description": "Filter by House name (dail or seanad).\nUsing an empty string retrieves results for both Houses.",
      "required": false,
      "type": "string",
      "default": "",
      "enum": [
        "",
        "dail",
        "seanad"
      ]
    },
    "chamberID": {
      "name": "chamber_id",
      "in": "query",
      "description": "Filter by house or committee uri, e.g. /ie/oireachtas/house/dail/32",
      "required": false,
      "type": "array",
      "collectionFormat": "csv",
      "allowEmptyValue": true,
      "default": []
    },
    "dateStart": {
      "name": "date_start",
      "in": "query",
      "description": "In the format YYYY-MM-DD (e.g. 1900-12-31)",
      "required": false,
      "type": "string",
      "format": "date"
    },
    "dateEnd": {
      "name": "date_end",
      "in": "query",
      "description": "In the format YYYY-MM-DD (e.g. 2025-12-31)",
      "required": false,
      "type": "string",
      "format": "date"
    },
...
    "skip": {
      "name": "skip",
      "in": "query",
      "description": "This allows skipping of records by a specific integer.",
      "required": false,
      "type": "integer",
      "default": 0
    },
    "limit": {
      "name": "limit",
      "in": "query",
      "description": "This allows the limiting of records to a specific integer.",
      "required": false,
      "type": "integer",
      "default": 50
    },
    "houseNo": {
      "name": "house_no",
      "in": "query",
      "description": "Filter by house number (integer only)",
      "required": false,
      "type": "integer"
    },
...
    "questionID": {
      "name": "question_id",
      "in": "query",
      "description": "Identifier for a Single Question, e.g. /ie/oireachtas/question/2025-04-10/pq_1",
      "required": false,
      "type": "string"
    },
    "questionNo": {
      "name": "question_no",
      "in": "query",
      "description": "Filter by question number",
      "required": false,
      "type": "integer"
    },
    "partyCode": {
      "name": "party_code",
      "in": "query",
      "description": "Filter by party code, e.g. Labour_Party",
      "required": false,
      "type": "string"
    },
    "partyID": {
      "name": "party_id",
      "in": "query",
      "description": "Filter by unique party identifier (uri), e.g. /ie/oireachtas/party/dail/32/Labour_Party",
      "required": false,
      "type": "string"
    },
    "constituencyCode": {
      "name": "const_code",
      "in": "query",
      "description": "Filter by constituency code, e.g. Administrative-Panel",
      "required": false,
      "type": "string"
    },
    "constituencyID": {
      "name": "const_id",
      "in": "query",
      "description": "Filter by unique constituency identifier (uri), e.g. /ie/oireachtas/house/seanad/26/panel/Administrative-Panel",
      "required": false,
      "type": "string"
    },
    "fuzzyNameSearch": {
      "name": "fuzzy_name_search",
      "in": "query",
      "description": "Filter by name with fuzzy search (typo safe)",
      "required": false,
      "type": "string"
    },
...
  }
}